// src/common/enums/entity.enum.ts
export enum EntityName {
    User = "user",
    Profile = "profile",
    Otp = "otp",
    category = "category",
    blog = "blog",
    blogCategory = "blogCategory",
    blogLikes = "blogLike",
    blogBookmarks = "blogBookmark",
    blogComments = "blogComment",
    image = "image",
    follow = "follow",
}